+++ app/src/main/kotlin/ac/mdiq/podcini/receiver/SPAReceiver.kt
@@ -36,7 +36,7 @@ class SPAReceiver : BroadcastReceiver() {
         for (url in feedUrls) {
             val feed = Feed(url, null, "Unknown podcast")
             feed.episodes.clear()
-            updateFeedFull(context, feed, false)
+            updateFeedFull(context, feed, removeUnlistedItems = false)
         }
          Logt(TAG, context.getString(R.string.sp_apps_importing_feeds_msg))
         runOnce(context)